pyproject.toml
[build-system]
build-backend = 'setuptools.build_meta'
requires = ['setuptools >= 44', 'wheel >= 0.37']

[tool.black]
line-length = 120

[project]
name = "robolint"
description = "Robolint linter"
version = "0.0.1"
dependencies = [
    "pylint[spelling]==2.17.4",
    "pytest==7.4.0",
    "stdlib_utils==0.4.8",
    "defusedxml==0.7.1",
    "overrides==7.3.1",
    "pyhumps==3.8.0",
    "lxml==4.9.3",
    "pyparsing==3.0.9",
]

[project.optional-dependencies]
dev = [
    "pre-commit==3.3.2",
    "pytest-randomly==3.12.0",
    "pytest-mock==3.11.1",
    "pytest-pylint==0.19.0",
    "pytest-cov==4.1.0",
    "mypy==1.4.0",
    "mock==5.1.0",
    "toml==0.10.2",
    "lxml-stubs==0.4.0",
]

[project.scripts]
enforce-workspace-settings = "robolint.hooks.enforce_workspace_settings:main"
clear-workspace-variables = "robolint.hooks.strip_workspace_config_values:main"
robolint = "robolint.run:run_pylint"
